Add files via upload
This Python code demonstrates a Convolutional Neural Network (CNN) model built using the Keras library for recognizing handwritten digits from the MNIST dataset. The model achieves high accuracy in classifying digits. It saves the trained model for future use. The GUI application allows users to draw digits on a canvas, and upon clicking the "Recognise" button, the model predicts the digit drawn with confidence percentage displayed.

Features:
1. CNN model with multiple layers for digit recognition.
2. GUI application using tkinter for drawing digits.
3. Ability to classify and display the predicted digit along with the confidence level.

Libraries used:
1. `keras`: For building and training the CNN model.
2. `tkinter`: For creating the graphical user interface.
3. `PIL`: For image processing and resizing.
4. `ImageGrab`: For capturing screen content to obtain drawn digit image.
5. `numpy`: For handling numerical arrays and data reshaping.

Learnings:
The code taught us how to use Keras to build and train a CNN model for digit recognition. It also demonstrated the implementation of a simple GUI using tkinter, allowing users to interactively draw digits and see the model's predictions in real-time. Additionally, we learned to capture the drawn digit image using `ImageGrab` and preprocess it for model inference.